优化代码结构
