+++ config/router.config.js
@@ -72,70 +72,72 @@ export default [
         ]
       },
       {
-        path: 'system-management',
-        name: 'system-management',
+        path: 'permission-management',
+        name: 'permission-management',
         icon: 'table',
         routes: [
           {
             authority: ['super_admin', 'user_page'],
-            path: '/system-management/user/list',
+            path: '/permission-management/user/list',
             name: 'user-list',
-            component: './system-management/user/list/Index',
+            component: './permission-management/user/list/Index',
           },
           {
             authority: ['super_admin', 'user_add'],
             hideInMenu: true,
-            path: '/system-management/user/add',
+            path: '/permission-management/user/add',
             name: 'user-add',
-            component: './system-management/user/add/Index',
+            component: './permission-management/user/add/Index',
           },
           {
             authority: ['super_admin', 'user_edit'],
             hideInMenu: true,
-            path: '/system-management/user/edit/:id',
+            path: '/permission-management/user/edit/:id',
             name: 'user-edit',
-            component: './system-management/user/edit/Index',
+            component: './permission-management/user/edit/Index',
           },
           {
             authority: ['super_admin', 'user_role_auth'],
             hideInMenu: true,
-            path: '/system-management/user/roleAuth/:userId',
+            path: '/permission-management/user/roleAuth/:userId',
             name: 'user-roleAuth',
-            component: './system-management/user/roleAuth/Index',
+            component: './permission-management/user/roleAuth/Index',
           },
           {
             // authority:['role_list'],
-            path: '/system-management/role/list',
+            path: '/permission-management/role/list',
             name: 'role-list',
-            component: './system-management/role/list/Index',
+            component: './permission-management/role/list/Index',
           },
           {
             // authority:['role_add'],
             hideInMenu: true,
-            path: '/system-management/role/add',
+            path: '/permission-management/role/add',
             name: 'role-add',
-            component: './system-management/role/add/Index',
+            component: './permission-management/role/add/Index',
           },
           {
             // authority:['role_edit'],
             hideInMenu: true,
-            path: '/system-management/role/edit/:id',
+            path: '/permission-management/role/edit/:id',
             name: 'role-edit',
-            component: './system-management/role/edit/Index',
+            component: './permission-management/role/edit/Index',
           },
           {
             // authority:['role_permissAuth'],
             hideInMenu: true,
-            path: '/system-management/role/permissAuth/:roleId',
+            path: '/permission-management/role/permissAuth/:roleId',
             name: 'role-permissAuth',
-            component: './system-management/role/permissAuth/Index',
+            component: './permission-management/role/permissAuth/Index',
           },
-          {
-            // authority:['permission_list'],
-            path: '/system-management/permission/list',
-            name: 'permission-list',
-            component: './system-management/permission/list/Index',
-          },
+
+        ]
+      },
+      {
+        path: 'system-management',
+        name: 'system-management',
+        icon: 'table',
+        routes: [
           {
             // authority:['menu_list'],
             path: '/system-management/menu/list',
@@ -155,6 +157,12 @@ export default [
             path: '/system-management/menu/edit/:id',
             name: 'menu-edit',
             component: './system-management/menu/edit/Index',
+          },
+          {
+            // authority:['permission_list'],
+            path: '/system-management/permission/list',
+            name: 'system-list',
+            component: './system-management/permission/list/Index',
           },
           {
             // authority:['dict_list'],
