+++ src/locales/zh-CN/menu.js
@@ -10,15 +10,16 @@ export default {
   'menu.basic-management.employee-list': '员工管理',
   'menu.basic-management.employee-add': '员工新增',
   'menu.basic-management.employee-edit': '员工编辑',
+  'menu.permission-management': '权限管理',
+  'menu.permission-management.user-list': '用户管理',
+  'menu.permission-management.user-add': '用户新增',
+  'menu.permission-management.user-edit': '用户编辑',
+  'menu.permission-management.user-roleAuth': '用户角色分配',
+  'menu.permission-management.role-list': '角色管理',
+  'menu.permission-management.role-add': '角色新增',
+  'menu.permission-management.role-edit': '角色编辑',
+  'menu.permission-management.role-permissAuth': '角色授权',
   'menu.system-management': '系统管理',
-  'menu.system-management.user-list': '用户管理',
-  'menu.system-management.user-add': '用户新增',
-  'menu.system-management.user-edit': '用户编辑',
-  'menu.system-management.user-roleAuth': '用户角色分配',
-  'menu.system-management.role-list': '角色管理',
-  'menu.system-management.role-add': '角色新增',
-  'menu.system-management.role-edit': '角色编辑',
-  'menu.system-management.role-permissAuth': '权限配置',
   'menu.system-management.permission-list': '功能管理',
   'menu.system-management.menu-list': '菜单管理',
   'menu.system-management.menu-add': '菜单新增',
